remove hardware specific rules
These two rules only applied to one of my machines. The rule for
the perc6i card is one which *should* not have been applied in
general, because you want to known about disk events.

(Although I recommend to use something which repeatedly checks
and notifies for that use case)